added post routes to blogsController

+++ src/modules/blogs/blogs.service.ts
@@ -13,9 +13,13 @@ export default class BlogsService {
     return this.repo.create(newBlog);
   }
   public async update(id: string, data: BlogInputModel): Promise<boolean> {
+    const blog = this.repo.get(id);
+    if (!blog) return false;
     return this.repo.update(id, data);
   }
   public async delete(id: string): Promise<boolean> {
+    const blog = this.repo.get(id);
+    if (!blog) return false;
     return this.repo.delete(id);
   }
 }
